requirements.txt
# ── Core ML / Retrieval ───────────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# ── PDF Parsing ────────────────────────────────────────────────────────────────
pymupdf>=1.24.0

# ── API Server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── LLM — OpenRouter is the primary provider (uses openai SDK) ───────────────
# OpenRouter: one key, 100+ models, free tiers available
# https://openrouter.ai/keys
openai>=1.30.0           # used for OpenRouter AND direct OpenAI

# Optional direct providers (only install what you need)
anthropic>=0.28.0        # direct Anthropic API
groq>=0.9.0              # direct Groq API (llama, fast + free)

# ── Numeric / Utilities ────────────────────────────────────────────────────────
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0     # loads .env file automatically

# ── Dev / Testing ──────────────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0
